Version 13.6

+++ vistas/Backoffice/eliminar.php
@@ -23,47 +23,47 @@ if (isset($_GET['id_camionero'])) {
 } else if (isset($_GET['id_almacen_cliente'])) {
     $id_almacen_cliente = $_GET['id_almacen_cliente'];
 
-    $paquetes = [];
-    $instruccion = "select * from almacena where id_almacen_cliente=$id_almacen_cliente";
-    $resultado = mysqli_query($conexion, $instruccion);
-    while ($fila = mysqli_fetch_assoc($resultado)) {
-        array_push($paquetes, $fila);
-    }
-    
-    if (isset($paquetes)){
-        if (count($paquetes) > 0) {
-           $instruccion = "delete from almacena where id_almacen_cliente=$id_almacen_cliente";
-           $conexion->query($instruccion);
-        }
-    }
-
-    $instruccion = "delete from tiene where id_almacen_cliente=$id_almacen_cliente";
+    $instruccion = "update almacen_cliente set estado='De baja' where id_almacen_cliente=$id_almacen_cliente";
     $conexion->query($instruccion);
-
-    $instruccion = "delete from almacen_cliente where id_almacen_cliente=$id_almacen_cliente";
-    $conexion->query($instruccion);
-    header("Location: op-almacen-cliente.php");
+    $respuesta = [
+        'error' => "Éxito",
+        'respuesta' => "Almacén eliminado"
+    ];
+    $conexion->close();
+    $respuesta = json_encode($respuesta);
+    header("Location: op-almacen-cliente.php?datos=" . urlencode($respuesta));
 } else if (isset($_GET['id_almacen_central'])) {
     $id_almacen_central = $_GET['id_almacen_central'];
 
 
-    $instruccion = "delete from almacen_central where id_almacen_central=$id_almacen_central";
+    $instruccion = "update almacen_central set estado='De baja' where id_almacen_central=$id_almacen_central";
     $conexion->query($instruccion);
-    header("Location: op-almacen-central.php");
+    $respuesta = [
+        'error' => "Éxito",
+        'respuesta' => "Almacén eliminado"
+    ];
+    $conexion->close();
+    $respuesta = json_encode($respuesta);
+    header("Location: op-almacen-central.php?datos=" . urlencode($respuesta));
 } else if (isset($_GET['id_plataforma'])) {
     $id_plataforma = $_GET['id_plataforma'];
 
 
-    $instruccion = "delete from plataforma where id_plataforma=$id_plataforma";
+    $instruccion = "update plataforma set estado='De baja' where id_plataforma=$id_plataforma";
     $conexion->query($instruccion);
-    header("Location: op-plataforma.php");
+    $respuesta = [
+        'error' => "Éxito",
+        'respuesta' => "Plataforma eliminada"
+    ];
+    $conexion->close();
+    $respuesta = json_encode($respuesta);
+    header("Location: op-plataforma.php?datos=" . urlencode($respuesta));
 } else if (isset($_GET['id_camion'])) {
     $id_camion = $_GET['id_camion'];
 
 
-    $instruccion = "delete from camion where id_camion=$id_camion";
-    $conexion->query($instruccion);
-    $instruccion = "delete from vehiculo where id_vehiculo=$id_camion";
+
+    $instruccion = "update vehiculo set estado='De baja' where id_vehiculo=$id_camion";
     $conexion->query($instruccion);
 
     $respuesta = [
@@ -76,11 +76,7 @@ if (isset($_GET['id_camionero'])) {
 } else if (isset($_GET['id_camioneta'])) {
     $id_camioneta = $_GET['id_camioneta'];
 
-
-    $instruccion = "delete from camioneta where id_camioneta=$id_camioneta";
-    $conexion->query($instruccion);
-
-    $instruccion = "delete from vehiculo where id_vehiculo=$id_camioneta";
+    $instruccion = "update vehiculo set estado='De baja' where id_vehiculo=$id_camioneta";
     $conexion->query($instruccion);
 
     header("Location: op-camionetas.php");
@@ -92,35 +88,30 @@ if (isset($_GET['id_camionero'])) {
     $conexion->query($instruccion);
     header("Location: op-ruta.php");
 
-} else if (isset($_GET['nom_usu'])) {
-    $nom_usu = $_GET['nom_usu'];
-
-
-    $instruccion = "delete from login where nom_usu='$nom_usu'";
-    $conexion->query($instruccion);
-    header("Location: op-usuarios.php");
-
 } else if (isset($_GET['id_empresa_cliente'])) {
 
     $id_empresa = $_GET['id_empresa_cliente'];
+
+    $instruccion = "update empresa_cliente set estado='De baja' where id_empresa_cliente=$id_empresa";
+    $conexion->query($instruccion);
     
+    $almacenes_cliente = [];
     $instruccion = "select * from tiene where id_empresa_cliente=$id_empresa";
     $resultado = mysqli_query($conexion, $instruccion);
-    $fila =  mysqli_fetch_assoc($resultado);
-    if (isset($fila)){
-        $respuesta = [
-            'error' => "Error",
-            'respuesta' => "Esta empresa tiene almacenes registrados"
-        ];
-    } else {
-        $instruccion = "delete from empresa_cliente where id_empresa_cliente=$id_empresa";
+    while ($row = mysqli_fetch_assoc($resultado)) {
+        array_push($almacenes_cliente, $row);
+    }
+    foreach ($almacenes_cliente as $almacen_cliente){
+        $id_almacen_cliente = $almacen_cliente["id_almacen_cliente"];
+        $instruccion = "update almacen_cliente set estado='De baja' where id_almacen_cliente=$id_almacen_cliente";
         $conexion->query($instruccion);
-        $respuesta = [
-            'error' => "Exito",
-            'respuesta' => "Empresa eliminada"
-        ];
     }
 
+    $respuesta = [
+        'error' => "Éxito",
+        'respuesta' => "Empresa eliminada"
+    ];
+    $conexion->close();
     $respuesta = json_encode($respuesta);
     header("Location: op-empresas-cliente.php?datos=" . urlencode($respuesta));
 
@@ -128,10 +119,30 @@ if (isset($_GET['id_camionero'])) {
     $id_camioneta_horario = $_GET['icth'];
     $fecha_salida = $_GET["fs"];
     $almacen_central_salida = $_GET["acs"];
+    date_default_timezone_set('America/Montevideo');
 
-    $instruccion = "delete from recoge where id_camioneta='$id_camioneta_horario' AND fecha_salida='$fecha_salida' AND almacen_central_salida='$almacen_central_salida'";
-    $conexion->query($instruccion);
-    header("Location: op-gestion-paquete-recogida.php");
+    
+    $fecha_hora_actual = time();
+    $fecha_salida1 = strtotime($fecha_salida);
+   
+
+    if ($fecha_salida1 > $fecha_hora_actual){
+        $instruccion = "delete from recoge where id_camioneta='$id_camioneta_horario' AND fecha_salida='$fecha_salida' AND almacen_central_salida='$almacen_central_salida'";
+        $conexion->query($instruccion);
+        $respuesta = [
+            'error' => "Éxito",
+            'respuesta' => "Horario eliminado"
+        ];
+    } else {
+        $respuesta = [
+            'error' => "Error",
+            'respuesta' => "No se puede eliminar un horario que ya pasó"
+        ];
+    }
+
+    $conexion->close();
+    $respuesta = json_encode($respuesta);
+    header("Location: op-gestion-paquete-recogida.php?datos=" . urlencode($respuesta));
 
 } else if (isset($_GET['id_camion_horario'])) {
     $id_camion_horario = $_GET['id_camion_horario'];
@@ -148,6 +159,13 @@ if (isset($_GET['id_camionero'])) {
         $instruccion = "delete from maneja where id_maneja=$id_maneja";
         $conexion->query($instruccion);
         header("Location: op-camionero-vehiculo.php");
+
+} else if (isset($_GET['id_usuario'])) {
+    $id_usuario = $_GET['id_usuario'];
+
+    $instruccion = "delete from login where id_usuario=$id_usuario";
+    $conexion->query($instruccion);
+    header("Location: op-usuarios.php");
 
 } else if (isset($_POST["todo"])){
